Single URL for the “unlock wallet” screen and the chat window
