+++ src/components/RoomButton.tsx
@@ -8,9 +8,9 @@ import { useWalletAccount, useWalletClient } from '$/features/wallet/hooks'
 import useIntercept from '$/hooks/useIntercept'
 import useJustInvited from '$/hooks/useJustInvited'
 import useRecentMessage from '$/hooks/useRecentMessage'
-import Avatar from '../../Avatar'
-import SidebarButton from '../../SidebarButton'
-import Text from '../../Text'
+import Avatar from './Avatar'
+import SidebarButton from './SidebarButton'
+import Text from './Text'
 import { PreferencesAction } from '$/features/preferences'
 import useIsRoomVisible from '$/hooks/useIsRoomVisible'
 import useIsRoomPinned from '$/hooks/useIsRoomPinned'
